Moved the view modifier in separate file.
